% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTT-package.R
\docType{package}
\name{PTT-package}
\alias{PTT}
\alias{PTT-package}
\title{PTT: Pólya Tree Type Models}
\description{
Bayesian nonparametric density and conditional-density estimation with
adaptive, Markov adaptive, and optional Pólya trees.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MaStatLab/PTT}
  \item Report bugs at \url{https://github.com/MaStatLab/PTT/issues}
}

}
\author{
\strong{Maintainer}: Li Ma \email{mastatlab@gmail.com}

Authors:
\itemize{
  \item Li Ma \email{mastatlab@gmail.com}
}

}
\keyword{internal}
